Move archive store util functions under utils direcotry
