+++ packages/beacon-node/src/chain/archiveStore/utils/archiveBlocks.ts
