+++ packages/beacon-node/src/chain/archiveStore/utils/pruneHistory.ts
