+++ packages/beacon-node/test/unit/chain/archive/nonCheckpoint.test.ts
@@ -1,7 +1,7 @@
 import {SLOTS_PER_EPOCH} from "@lodestar/params";
 import {Slot} from "@lodestar/types";
 import {beforeAll, describe, expect, it} from "vitest";
-import {getNonCheckpointBlocks} from "../../../../src/chain/archiverStore/archiveBlocks.js";
+import {getNonCheckpointBlocks} from "../../../../src/chain/archiverStore/utils/archiveBlocks.js";
 
 describe("chain / archive / getNonCheckpointBlocks", () => {
   beforeAll(() => {
